requirements.txt
# WOLF_AI Dependencies

# ==================== Core ====================
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# ==================== HTTP/API ====================
httpx>=0.25.0
requests>=2.31.0

# ==================== WebSocket ====================
websockets>=12.0

# ==================== Telegram Bot ====================
python-telegram-bot>=20.0

# ==================== Configuration ====================
python-dotenv>=1.0.0

# ==================== CLI/UI ====================
rich>=13.0.0

# ==================== Memory (optional) ====================
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# ==================== Voice Control (optional) ====================
# SpeechRecognition>=3.10.0
# pyttsx3>=2.90
# pyaudio>=0.2.13  # Required for microphone input

# ==================== AI Models (optional) ====================
openai>=1.0.0  # For ChatGPT integration
anthropic>=0.40.0  # For Claude API (Alpha brain)
# google-generativeai>=0.3.0
